libc: match ioctl request type to contrib type
There is a type mismatch as in the FreeBSD contrib code the type of the
request is 'unsigned long'. So far, only I/O controls where the request
falls into the signed range where used and this was not a problem.

Some of the SNDCTL requests, however, have the bit set.

Fixes #3887.

+++ repos/libports/include/libc-plugin/plugin.h
@@ -110,7 +110,7 @@ namespace Libc {
 			virtual int getsockopt(File_descriptor *, int level,
 			                       int optname, void *optval,
 			                       socklen_t *optlen);
-			virtual int ioctl(File_descriptor *, int request, char *argp);
+			virtual int ioctl(File_descriptor *, unsigned long request, char *argp);
 			virtual int listen(File_descriptor *, int backlog);
 			virtual ::off_t lseek(File_descriptor *, ::off_t offset, int whence);
 			virtual int mkdir(const char *pathname, mode_t mode);
