+++ repos/libports/src/lib/libc/internal/vfs_plugin.h
@@ -91,7 +91,7 @@ class Libc::Vfs_plugin : public Plugin
 		 */
 		void _vfs_write_mtime(Vfs::Vfs_handle&);
 
-		int _legacy_ioctl(File_descriptor *, int , char *);
+		int _legacy_ioctl(File_descriptor *, unsigned long, char *);
 
 		/**
 		 * Call functor 'fn' with ioctl info for the given file descriptor 'fd'
@@ -177,7 +177,7 @@ class Libc::Vfs_plugin : public Plugin
 		int     fsync(File_descriptor *fd) override;
 		int     ftruncate(File_descriptor *, ::off_t) override;
 		ssize_t getdirentries(File_descriptor *, char *, ::size_t , ::off_t *) override;
-		int     ioctl(File_descriptor *, int , char *) override;
+		int     ioctl(File_descriptor *, unsigned long, char *) override;
 		::off_t lseek(File_descriptor *fd, ::off_t offset, int whence) override;
 		int     mkdir(const char *, mode_t) override;
 		File_descriptor *open(const char *path, int flags) override;
